Added better tarball checking
Lets the user know which tarball they need. Renamed jailnumber, and
added a rootfolder for the jail's location. Additionally, if the
rootfolder is not specified, then the the default location (within
thebrig's root) is used, and folder created if need be. Also, added a
jail renumbering algorithm in case of jailno conflict.