Trailing slashes repair
The symptom was that upon first arriving at the rudimentary config page,
clicking save would result in the installed location to have its
contents deleted, and an errorneous folder created. This was due to
creating a directory {$new_location}work, as if there was a trailing
slash on the newlocation. However, this was not always put their
correctly, so _config had to be modified to properly guarantee a
trailing slash on the rootfolder's name.