Fixed processes
Now thebrig_check correctly determines running and sleeping processes the same way that top does. However, this version will work, whether or not the jail has the devfs mount enabled. The previous version of this code did not realize that processes with a state of S+J were actually running. This code ammends that - processes that are S* || I* && !S+* are now considered sleeping.