Massive Update
This enables the creation of a jail from either tarballs or from a
template jail. Also, it implements slim jails. This required a a slight
change to the layout of the config array. If jails have problems, then
just recreate them. I also added shutting down/unmounting a jail before
backing it up. I also added code to remove the old fstab entries. THe
last thing was to re-organize the layout of the tools pages.