Cron Scripts
These php scripts are called as part of the cronjob to update either the
freebsd-upate patch or the ports tree. This is done because of the
binaries that need to be temporarily moved around to allow for these two
operations. Previously, we would have had to move those files, then let
either of the original sbin scripts wait up to an hour (randomly), and
then carry out the updates, and then the temp files would be restored
following it. Now, however, the random wait (to ease cronjob loads on
servers), is carried out in the php script, PRIOR to temporarily moving
the files. This ensures that the other binaries are in place for only a
minimum of time.