Continued Ports debugging
The check to see if a particular jail should have access to the ports
tree is not to see if the "ports" key is defined at all. If it is, then
we carry out some things. The biggest noticeable change is in the proper
display of checked boxes to show the current status. The status page is
correct, the mounting and unounting is also, and the fstab is populated
correctly.