Relocated sleep, syntax fixes and logic fix
Sleep moved outside if statement.

The location to store files was not being declared in the script (it
works just fine on the website). This meant the basejail's storage
location was now being made available to the freebsd-update script.

Line 24 had an open curly brace but no closure.

Line 44 evaluated an erroneous variable to test if the template jail had
been filled. Instead, the presence of the /usr directory is used to
check if the template jail should have the update tasks carried out.