Fix erroneous port activation
The default creation and subsequent process for ANY jail would add the
"ports" key to every jail, whether or not it was meant to be. This fix
at line 348 fixes what happens when you edit a jail, as does the 525
stuff. The hidden values are used for items that don't have html
elements to return them in POST.